navbar optimizations
