--- pages/screens.js
@@ -1,18 +1,0 @@
-import Head from 'next/head';
-import styles from '../styles/Home.module.css';
-
-export default function Screens () {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Branda - Screens</title>
-      </Head>
-
-      <main className={styles.main}>
-      </main>
-
-      <footer className={styles.footer}>
-      </footer>
-    </div>
-  );
-}
